requirements.txt
python-telegram-bot[job-queue]==20.7
openai==1.12.0
python-dotenv==1.0.0
aiohttp==3.9.3
Pillow==10.2.0
pytz==2024.1